Fixed RR throughput.

+++ Assignment 1/Assignment 1/main.c
@@ -40,9 +40,9 @@ void print_statistics()
 {
     
     int throughput = globalStatistics.throughput / NUM_TRIALS;
-    float avgTurnaround = globalStatistics.turnaround_time / NUM_TRIALS;
-    float avgWaitTime = globalStatistics.waiting_time / NUM_TRIALS;
-    float avgResponse = globalStatistics.response_time / NUM_TRIALS;
+    float avgTurnaround = globalStatistics.turnaround_time / (float)NUM_TRIALS;
+    float avgWaitTime = globalStatistics.waiting_time / (float)NUM_TRIALS;
+    float avgResponse = globalStatistics.response_time / (float)NUM_TRIALS;
     printf("\nAvg. Turnaround Time: %f\nAvg. Wait Time: %f\nAvg. Response Time: %f\nAvg. Throughput: %d\n\n",
            avgTurnaround, avgWaitTime, avgResponse, throughput);
 }
